Fix table detection in querylog
